+++ .dumirc.ts
@@ -2,6 +2,11 @@ import { defineConfig } from 'dumi'
 
 export default defineConfig({
     outputPath: 'docs-dist',
+    // apiParser: {},
+    // resolve: {
+    //   // 配置入口文件路径，API 解析将从这里开始
+    //   entryFile: './src/index.ts',
+    // },
     themeConfig: {
         name: 'FOundUI',
         nav: [
