+++ src/index.ts
@@ -1,4 +1,8 @@
 import { Print } from 'FOundUI'
+/**
+ * @version 1.1.0
+ *
+ */
 Print('FOundUI', '1.0.0')
 
 export { default as Button } from './Button'
